implement react router dom to separate default login page from register page
